requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
flower==1.0.0
gunicorn==20.1.0
humanize==3.11.0
kombu==5.1.0
Markdown==3.3.4
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
